Somehow in this mess the marker problem got fixed
Refactor pdg construction. Split away async support and calling conventions
Flowistry now collects info about a place and passes it on
Add test case for async_trait
Hash and Eq for DepNode now ignores unimportant fields
