+++ crates/flowistry_pdg/Cargo.toml
@@ -12,4 +12,6 @@ rustc = []
 [dependencies]
 cfg-if = "1.0.0"
 internment = { version = "0.7.4", features = ["serde"] }
-serde = { version = "1.0.193", features = ["derive"] }
+
+strum = { workspace = true }
+serde = { workspace = true }
